Get information about Grand Exchange prices of items in OSRS

Use the Weird Gloop API to fetch information about Grand Exchange prices in Oldschool Runescape

Usage

osrs_api(items, history = "latest", user_agent = "GE_price_tracker")

Arguments

items

character vector of the items

history

character with type of request. Default is latest, but all is also available

user_agent

character of the user agent

Value

A list of class osrs_api

Examples

# Get the latest prices of both a bucket and the abyssal whip
bucket_and_whip <- osrs_api(items = c("bucket", "Abyssal whip"), history = "latest")

# Get the entire price history of the abyssal whip
whip <- osrs_api(items = "Abyssal_whip", history = "all")