Create a tibble of information with prices of requested item

The function is a wrapper of osrs_api() that returns a tibble which is easier to work with than the list object from the osrs_api function.

Usage

create_item_data(items, history = "all", user_agent = "GE_price_tracker")

Arguments

items

character vector of the items

history

character with type of request. Default is latest, but all is also available

user_agent

character of the user agent

Value

A tibble with information about items, prices, dates

Examples

# Create a data set with information about the item
create_item_data(items = "Abyssal_whip", history = "all")
#> # A tibble: 3,247 × 6
#>    item         id      price rel_day date       volume
#>    <chr>        <chr>   <int>   <dbl> <date>      <int>
#>  1 Abyssal whip 4151  1830704       0 2024-04-23   4730
#>  2 Abyssal whip 4151  1844101       1 2024-04-22   5006
#>  3 Abyssal whip 4151  1853530       2 2024-04-21  10618
#>  4 Abyssal whip 4151  1875130       3 2024-04-20   6541
#>  5 Abyssal whip 4151  1899375       4 2024-04-19   6332
#>  6 Abyssal whip 4151  1901588       5 2024-04-18   3676
#>  7 Abyssal whip 4151  1924803       6 2024-04-17   5471
#>  8 Abyssal whip 4151  1959491       7 2024-04-16   6509
#>  9 Abyssal whip 4151  1976147       8 2024-04-15  13050
#> 10 Abyssal whip 4151  1999665       9 2024-04-14   8021
#> # ℹ 3,237 more rows

# Same can be done for other values of history
create_item_data(items = c("bucket", "Abyssal whip"), history = "latest") |>
create_item_data(items = c("Dragon dagger"), history = "sample")
#> Error in purrr::map(items, osrs_api, history = history, user_agent = user_agent): ℹ In index: 1.
#> Caused by error in `httr::user_agent()`:
#> ! is.character(agent) is not TRUE